+++ test/iterable/chunked_test.dart
@@ -7,6 +7,7 @@ void main() {
     expect(() => [].chunked(-1), throwsArgumentError);
     expect(() => [].chunked(0), throwsArgumentError);
     expect([].chunked(2).toList(), equals([]));
+
     expect(
         [1, 2, 3, 4, 5, 6].chunked(2).toList(),
         equals([
@@ -14,6 +15,7 @@ void main() {
           [3, 4],
           [5, 6]
         ]));
+
     expect(
         [1, 2, 3, 4, 5].chunked(2).toList(),
         equals([
@@ -21,11 +23,29 @@ void main() {
           [3, 4],
           [5]
         ]));
+
     expect(
         [1, 2, 3, 4, 5, 6].chunked(4).toList(),
         equals([
           [1, 2, 3, 4],
           [5, 6]
+        ]));
+  });
+
+  test("iterable chunked with fill", () {
+    expect([].chunked(2, fill: () => 99).toList(), equals([]));
+
+    expect(
+        [1].chunked(2, fill: () => 99).toList(),
+        equals([
+          [1, 99]
+        ]));
+
+    expect(
+        [1, 2, 3, 4, 5].chunked(3, fill: () => 0).toList(),
+        equals([
+          [1, 2, 3],
+          [4, 5, 0],
         ]));
   });
 }