+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## next
+
+- Fixed typo in api documentation
+
 ## 1.0.0
 
 * Initial release