+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## next
 
-- Fixed typo in api documentation
+- Fixed typo in API documentation
+- More efficient implementation for `minBy` and `maxBy`
+- Iterable `chunked` method has additional parameter `fill`
 
 ## 1.0.0